+++ probes/notCreatedRecently/def.yml
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-id: notCreatedInLast90Days
+id: notCreatedRecently
 short: Checks that the project was not created in the last 90 days.
 motivation: >
   When Scorecard checks the activity of a project in the last 90 days, the project may not have been created before the last 90 days. As such, Scorecard cannot give an accurate score. This probe helps Scorecard assess whether it can give an accurrate score when checking the project activity in the last 90 days.
@@ -20,7 +20,7 @@ implementation: >
   The implementation checks the creation date is within the last 90 days.
 outcome:
   - If the project was created within the last 90 days, the outcome is OutcomeNegative (0).
-  - If the project was created before the last 90 days, the outcome is OutcomePositive (1).
+  - If the project was created before the last 90 days, the outcome is OutcomePositive (1). The finding will include a "lookBackDays" value which is the time period that the probe looks back in. 
 remediation:
   effort: Low
   text:
